Added failure tests
* Added tests for fails() and reverts()
* Restructured tests into two files
  * Event related
  * Failures related
