+++ package.json
@@ -36,6 +36,7 @@
   },
   "devDependencies": {
     "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "coveralls": "^3.0.2",
     "eslint": "^5.6.1",
     "eslint-config-airbnb-base": "^13.1.0",
