--- test/fails.test.js
@@ -1,8 +1,0 @@
-const { assert } = require('chai');
-const truffleAssert = require('..');
-
-describe('fails', () => {
-  it('should initialise', () => {
-    assert.isDefined(truffleAssert);
-  });
-});
